third_party/WebKit/LayoutTests/fast/forms/button/button-disabled-blur.html

<!DOCTYPE html>
<body>
<style>
button:disabled {
    color: #aaa;
}
button {
    -webkit-appearance: none;
}
</style>
<button autofocus onclick="this.disabled = true;">Button</button>
<script src="../../../resources/js-test.js"></script>
<script>
jsTestIsAsync = true;
var button = document.querySelector('button');
shouldBe('document.activeElement', 'button');
debug('Clicking a button makes the button disabled.');
window.onload = function() {
    button.addEventListener('blur', function() {
        testPassed('blur event was disaptched.');
        button.remove();
        finishJSTest();
    }, false);
    // Need to wait until CheckFocusedElementTask is unqueued.
    setTimeout(function() {
        button.click();
    }, 20);
    setTimeout(function() {
        testFailed('No blur event.');
        finishJSTest();
    }, 2000);
};
</script>
</body>
