third_party/blink/web_tests/fast/forms/button/button-disabled-blur-expected.txt

PASS document.activeElement is button
Clicking a button makes the button disabled.
PASS blur event was disaptched.
PASS successfullyParsed is true

TEST COMPLETE
